Pepper/pepper_DL/server/requirements.txt
numpy
pandas
Flask
opencv-python==4.2.0.32
Pillow
requests
jsonpickle
